Projectile, new-line-from-anywhere, copy-line-down.
new-line-from-anywhere for using M-RET to get an indented newline and
move to it regardless of cursor column in current line. copy-line-down
to replicate current line below it. use-package-always-ensure removes
need to put :ensure t for all use-package invocations. No longer using
web-mode for JavaScript because it messes up comments. By default
emacs doesn't delete current selection when typing over it without
deleting it first so delete-selection-mode is now active to override
this.